Important Changes.  Read Description of commit.
* DFError now prints itself in an easy to read manner.  Just call
`print(error);`(print is a method i made in database.DFDatabase because
I was tired of called System.out.print)

* A rudimentary queue system has been put in place.  It will enable
explicit running of code on the main thread.  I use it to make sure
your delegate stuff is called on the main thread always.  To use the
queue, place this line of code into your file:
`queue.add(() -> /*Insert Code to Run on Main Thread Here*/);`

* Fixed extraneous “Processing” printout